alarm/antilles-wechat-agent/requirements.txt
falcon>=1.3
PyYAML>=3.10
wechatpy[cryptography]>=1.4.0
setuptools>=36.0
jsonschema>=2.5.1